+++ os/kernel/port/line.cpp
@@ -23,7 +23,7 @@
 #include "line.h"
 
 Line::
-Line(es::Callback* callback, u8 bits, u8 channels, u8 rate) :
+Line(es::Callback* callback, u8 bits, u8 channels, u16 rate) :
     callback(callback),
     bits(bits),
     channels(channels),
@@ -171,7 +171,7 @@ release()
 }
 
 InputLine::
-InputLine(es::Callback* callback, u8 bits, u8 channels, u8 rate) :
+InputLine(es::Callback* callback, u8 bits, u8 channels, u16 rate) :
     Line(callback, bits, channels, rate)
 {
 }
@@ -212,7 +212,7 @@ write(const void* src, int count)
 }
 
 OutputLine::
-OutputLine(es::Callback* callback, u8 bits, u8 channels, u8 rate) :
+OutputLine(es::Callback* callback, u8 bits, u8 channels, u16 rate) :
     Line(callback, bits, channels, rate)
 {
 }
